fix: fix: bitInt
